Fix input IRQ logic, NTP calculations, and panel coordinate mapping (#1073)
- Refactored `kernel/drivers/input/input.c` to use `irq_save()` and `irq_restore(flags)` instead of raw inline assembly.
- Moved `#include "../../include/nvram.h"` to the top of `kernel/shell/cmd_devices.c` to prevent local scope pollution.
- Fixed the Y-coordinate condition in `panel_time()` inside `kernel/shell/cmd_panel.c` for proper mapping.
- Added explicit type casting in `cmd_ntp()` inside `kernel/shell/cmd_time.c` to properly perform 64-bit subtractions with `NTP_TIMESTAMP_DELTA`.

+++ kernel/shell/cmd_time.c
@@ -145,10 +145,10 @@ void cmd_ntp(const char* args) {
     time_t unix_time;
     if ((ntp_secs & 0x80000000) == 0) {
         // MSB is 0, NTP timestamp has rolled over to era 1 (starts in 2036)
-        unix_time = (time_t)ntp_secs + 0x100000000ull - NTP_TIMESTAMP_DELTA;
+        unix_time = (time_t)((uint64_t)ntp_secs + 0x100000000ull - NTP_TIMESTAMP_DELTA);
     } else {
         // Era 0
-        unix_time = (time_t)ntp_secs - NTP_TIMESTAMP_DELTA;
+        unix_time = (time_t)((uint64_t)ntp_secs - NTP_TIMESTAMP_DELTA);
     }
 
     rtc_time_t new_time;
